CPU version of the sourcecode
For the moment, changes are made during compilation.

I still need to find out more details concerning CUDA.
I think that it is possible to set up a CUDA installation without having the required GPU.

If such configuration can exist (or even for outdated CUDA versions), I need to try it in order to be able to deal with CUDA problems directly from C/C++ code.
Indeed, CUDA libraries should be loaded but running a kernel should raise an exception that needs to be handle C/C++ side.